code cleaning from pr feedback
