+++ package.json
@@ -42,8 +42,6 @@
   },
   "dependencies": {
     "concurrently": "^8.2.2",
-    "graphql": "^16.8.1",
-    "graphql-request": "^6.1.0",
     "prettier": "^3.0.3"
   },
   "pnpm": {
