+++ packages/grant-explorer/.env.example
@@ -68,4 +68,5 @@ REACT_APP_PASSPORT_AVALANCHE_API_KEY="test-key"
 REACT_APP_PASSPORT_API_COMMUNITY_ID="0000"
 REACT_APP_PASSPORT_API_COMMUNITY_ID_AVALANCHE="0000"
 
+# create key following https://www.opensource.observer/docs/get-started/
 REACT_APP_OSO_API_KEY=
