+++ packages/grant-explorer/package.json
@@ -76,6 +76,8 @@
     "ethereum-blockies": "^0.1.1",
     "ethers": "^5.7.2",
     "framer-motion": "^10.15.0",
+    "graphql": "^16.8.1",
+    "graphql-request": "^6.1.0",
     "history": "^5.3.0",
     "html-react-parser": "^3.0.7",
     "https-browserify": "^1.0.0",
