+++ packages/grant-explorer/src/features/common/utils/utils.ts
@@ -13,3 +13,18 @@ export const getFormattedRoundId = (roundId?: string | number): string => {
     return roundId;
   }
 };
+
+export function formatTimeAgo(dateString : number) {
+  const date = new Date(dateString);
+  const now = new Date();
+  const diffTime = Math.abs(now.getTime() - date.getTime()); // Difference in milliseconds
+  const diffMonths = Math.round(diffTime / (1000 * 60 * 60 * 24 * 30)); // Convert to months
+
+  if (diffMonths === 0) {
+    return 'This month';
+  } else if (diffMonths === 1) {
+    return 'Last month';
+  } else {
+    return `${diffMonths} months`;
+  }
+}
